Fix adding completed subtasks
Previously, if a completed subtask was added, it would not have its
parent set. Currently not only this is fixed but also the order of
completed tasks is preserved.

This comes with a cost of sending more request to the Google Tasks API.
It should be considered to make this as a batched request for the
completed tasks.